ConfigHelper renamed as ResourceReaderHelper Also moved a helper method from BootConfig to ResourceReaderHelper
